requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
openai
huggingface_hub
faiss-cpu
tiktoken